deepseek/requirements.txt
streamlit>=1.24.0
langchain>=0.0.267
langchain-huggingface>=0.0.6
langchain-community>=0.0.10
langchain-deepseek>=0.0.1
python-dotenv>=1.0.0
faiss-cpu>=1.7.4
transformers>=4.30.2
accelerate>=0.20.3
torch>=2.0.1
nest-asyncio>=1.5.6
pypdf>=3.15.1
pymupdf>=1.22.5
sentence-transformers>=2.2.2
numpy>=1.24.3
pydantic>=2.4.2
fastapi>=0.104.0
uvicorn>=0.23.2
python-multipart>=0.0.6 